% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump2pwa.R
\name{generate_manifest}
\alias{generate_manifest}
\title{Generate a manifest.json file for a PWA}
\usage{
generate_manifest(
  path = "manifest.json",
  name = "My PWA",
  short_name = "PWA",
  start_url = "/",
  display = "standalone",
  orientation = "any",
  background_color = "#ffffff",
  theme_color = "#ffffff",
  icons = list(list(src = "icon-192x192.png", sizes = "192x192", type = "image/png"),
    list(src = "icon-512x512.png", sizes = "512x512", type = "image/png"))
)
}
\arguments{
\item{path}{The path where the manifest.json file will be saved (default: "manifest.json").}

\item{name}{The name of the PWA.}

\item{short_name}{A shorter name for the PWA (for homescreen).}

\item{start_url}{The start URL of the PWA (usually the root "/").}

\item{display}{The display mode of the app (e.g., "standalone", "fullscreen", etc.).}

\item{orientation}{The screen orientation of the app (e.g., "any", "portrait", "landscape").}

\item{background_color}{The background color of the splash screen (default: "#ffffff").}

\item{theme_color}{The theme color of the app (default: "#ffffff").}

\item{icons}{A list of icons with their paths and sizes, used in the manifest file.}
}
\value{
Generates a manifest.json file at the specified path.
}
\description{
This function generates a manifest.json file for a Progressive Web App (PWA).
It includes metadata such as the app's name, icons, start URL, and display options.
}
\examples{
\donttest{
generate_manifest(
  path = "manifest.json",
  name = "My PWA",
  short_name = "PWA",
  start_url = "/",
  display = "standalone",
  orientation = "any",
  background_color = "#ffffff",
  theme_color = "#ffffff",
  icons = list(
    list(src = "icon-192x192.png", sizes = "192x192", type = "image/png"),
    list(src = "icon-512x512.png", sizes = "512x512", type = "image/png")
  )
)
}

}
