% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jump2pwa.R
\name{setup_pwa}
\alias{setup_pwa}
\title{Set up a PWA by generating manifest.json, service-worker.js, and an offline page}
\usage{
setup_pwa(
  pwa_name = "My PWA",
  short_name = "PWA",
  start_url = "/",
  display = "standalone",
  orientation = "any",
  background_color = "#ffffff",
  theme_color = "#ffffff",
  icons = list(list(src = "icon-192x192.png", sizes = "192x192", type = "image/png"),
    list(src = "icon-512x512.png", sizes = "512x512", type = "image/png")),
  cache_name = "my-pwa-cache-v1",
  assets = c("/", "/index.html", "/css/style.css"),
  caching_strategy = "cache-first",
  offline_message = "You are currently offline."
)
}
\arguments{
\item{pwa_name}{The name of your PWA.}

\item{short_name}{A short name for your PWA.}

\item{start_url}{The start URL of the PWA.}

\item{display}{The display mode of the app (default: "standalone").}

\item{orientation}{The screen orientation of the app (default: "any").}

\item{background_color}{The background color of the splash screen (default: "#ffffff").}

\item{theme_color}{The theme color of the app (default: "#ffffff").}

\item{icons}{A list of icons for the manifest.json.}

\item{cache_name}{The name of the cache in the service worker.}

\item{assets}{A list of assets to cache in the service worker.}

\item{caching_strategy}{The caching strategy to use ("cache-first", "network-first", etc.).}

\item{offline_message}{The message to display on the offline page.}
}
\value{
Generates all the required files for the PWA.
}
\description{
This function sets up a Progressive Web App (PWA) by generating the necessary manifest.json file,
service-worker.js file, and an offline page.
}
\examples{
setup_pwa(pwa_name = "My PWA", short_name = "PWA", start_url = "/", display = "standalone",
          orientation = "any", background_color = "#ffffff", theme_color = "#ffffff",
          icons = list(list(src = "icon-192x192.png", sizes = "192x192", type = "image/png"),
                       list(src = "icon-512x512.png", sizes = "512x512", type = "image/png")),
          cache_name = "my-pwa-cache-v1", assets = c("/", "/index.html", "/css/style.css"),
          caching_strategy = "cache-first", offline_message = "You are currently offline.")
}
